+++ tasks/tyre_impression_tasks.py
@@ -4,6 +4,6 @@ from services.tyre_impression_processing_service import TyreImpressionProcessing
 
 
 @shared_task(autoretry_for=(Exception,), retry_backoff=True, max_retries=3)
-def process_tyre_impression_task(tyre_impression: TyreImpression):
+def process_tyre_impression_task(tyre_impression_id: int):
     service = TyreImpressionProcessingService()
-    return service.process_tyre_impression(tyre_impression)
+    return service.process_tyre_impression(tyre_impression_id)